[#3]refactor: move type definitions to types directory
